+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- fix permission denied / `EACCES` error
+  when `$POSTMAP_PATHS` contains path to file not owned by root.
+  keep `postmap` from changing uid/gid to file owner
+  as user might not have write perms on dir
+  (e.g., due to mount with `dockerd --userns-remap`).
 
 ## [1.0.2] - 2019-12-21
 ### Fixed